[SPARK-17925][SQL] Break fileSourceInterfaces.scala into multiple pieces
